rename contribute to upload in the UI, but not in the implementation
