+++ src/cli/library_upload.js
@@ -40,25 +40,25 @@ export class CLILibraryContributeCommandSite extends LibraryContributeCommandSit
 	}
 
 	contributingLibrary(promise, library) {
-		return spin(promise, `Contributing library ${chalk.green(library.name)}`);
+		return spin(promise, `Uploading library ${chalk.green(library.name)}`);
 	}
 
 	contributeComplete(library) {
-		return log.success(`Library ${chalk.green(library.name)} was successfully contributed.\n` +
+		return log.success(`Library ${chalk.green(library.name)} was successfully uploaded.\n` +
 		`Add it to your project with ${chalk.green('particle library add ' + library.name)}`);
 	}
 }
 
 export default ({lib, factory, apiJS}) => {
-	factory.createCommand(lib, 'contribute', 'contributes a library', {
+	factory.createCommand(lib, 'upload', 'uploads a private version of a library', {
 		options: {
 			'dryRun': {
 				required: false,
 				boolean: true,
-				description: 'perform validation steps but don\'t actually contribute the library.'
+				description: 'perform validation steps but don\'t actually upload the library.'
 			}
 		},
-		handler: function LibraryContributeHandler(argv) {
+		handler: function LibraryUploadHandler(argv) {
 			const site = new CLILibraryContributeCommandSite(argv, process.cwd(), buildAPIClient(apiJS));
 			const cmd = new LibraryContributeCommand();
 			return site.run(cmd);
